pyproject.toml
[tool.poetry]
name = "django-getpaid-paynow"
version = "0.1.2"
description = "Django-GetPaid plugin for mBank payNow service."
authors = ["Dominik Kozaczko <dominik@kozaczko.info>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/django-getpaid/django-getpaid-paynow"
packages = [
    { include = "getpaid_paynow" }
]


[tool.poetry.dependencies]
python = "^3.6"
django-getpaid = "^2.0.0-rc.6"
requests = "^2.23.0"
swapper = "^1.1.2"


[tool.poetry.dev-dependencies]
pre-commit = "^2.1.1"
factory-boy = "^2.12.0"
black = "^19.10b0"
tox = "^3.14.5"
ipdb = "^0.13.2"
pytest = {version = "^5.3.5", optional = true}
codecov = {version = "^2.0.16", optional = true}
coverage = {version = "^5.0.3", optional = true}
requests-mock = {version = "^1.7.0", optional = true}
pytest-cov = {version = "^2.8.1", optional = true}
pytest-django = {version = "^3.8.0", optional = true}
sphinx = {version = "^2.4.3", optional = true}

[tool.poetry.extras]
docs = ["sphinx"]
test = ["pytest", "codecov", "coverage", "request-mock", "pytest-cov", "pytest-django"]

[tool.black]
line-length = 88
target-version = ['py36']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.cache
    | \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | __pycache__
    | _build
    | buck-out
    | build
    | dist
    | docs
    | migrations
  )/

)
'''


[tool.isort]
multi_line_output = 3
include_trailing_comma = false
line_length = 88
known_third_party = ["django", "factory", "getpaid", "orders", "requests", "swapper"]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
